requirements.txt
fastapi
uvicorn
tensorflow
numpy
pillow
google-cloud-storage